network: Encode information for systemd-networkd-wait-online
The network-online.target can use systemd-networkd-wait-online.service
to wait for all interfaces to come up. It will fail if the interfaces
didn't came up but sometimes it is actually ok for some interfaces to
be down because they are unused or they are just one of two parts of a
bond. We should encode when interfaces will never come up and when it
is acceptable to have interfaces in a degraded state and which.
Extend the network logic to handle this additional configuration. For
Packet we expect the metadata to give all interfaces and any other
physical NICs can be set to "unmanaged" so that we don't wait for them.
We also allow bonds to operate with only one working link, and we don't
wait for all bonded interfaces to be configured.